+++ module/actor/entity.js
@@ -415,6 +415,7 @@ export class ActorPF extends Actor {
 
   preUpdate(data) {
     data = flattenObject(data);
+    const fullData = mergeObject(this.data, expandObject(data));
 
     // Apply settings
     // Set used spellbook flags
@@ -517,6 +518,30 @@ export class ActorPF extends Actor {
       // Remove resource
       if (item == null || createTag(item.name) !== tag) {
         data[`data.resources.-=${tag}`] = null;
+      }
+    }
+
+    // Update spellbook slots
+    {
+      const slots = {};
+      for (let [sbKey, sb] of Object.entries(getProperty(this.data, "data.attributes.spells.spellbooks"))) {
+        for (let a = 0; a < 10; a++) {
+          setProperty(slots, `${sbKey}.${a}`, getProperty(fullData, `data.attributes.spells.spellbooks.${sbKey}.spells.spell${a}.max`) || 0);
+        }
+      }
+      const spells = this.items.filter(o => {
+        return o.type === "spell" && !getProperty(o.data, "data.domain");
+      });
+
+      for (let i of spells) {
+        const sb = i.spellbook;
+        if (!sb || (sb && sb.spontaneous)) continue;
+        const sbKey = i.data.data.spellbook;
+        const a = i.spellLevel;
+        let uses = getProperty(slots, `${sbKey}.${a}`);
+        uses -= i.maxCharges;
+        setProperty(slots, `${sbKey}.${a}`, uses)
+        data[`data.attributes.spells.spellbooks.${sbKey}.spells.spell${a}.value`] = uses;
       }
     }