+++ CHANGELOG.md
@@ -8,8 +8,9 @@
 
 ### Changelog
 
-- Prepared spellbooks now refresh spell slots on rest (#137)
-- Added an option to mark a spell as a domain or school bonus spell, which will not cost any spell slots on rest
+- Updating charges or maximum charges on an item such as a spell will update the other value of the two as well, depending on what changed
+- Spell slots of prepared spellbooks are now updated in real time (#137)
+- Added an option to mark a spell as a domain or school bonus spell, which will not cost any spell slots
 - Added a numeric field for spell material component costs
 
 ### Localization